refactor: apply reconciled nuclear-review findings
Whole-codebase maintainability pass (cross-model verified, net -1.2k lines)
plus the correctness fixes it surfaced. Behavior-preserving except the
explicit bug fixes below; no consumer-facing API change.

Correctness:
- shopifyFetch validates response payloads (opt-in dataSchema), not just
  the GraphQL envelope; Shopify README corrected to match.
- AddToCart sets disabled when no variant is selected (no no-op action).
- HubSpot form parser keeps every field in a multi-field group.
- WebGL fluid sim is frame-rate independent (delta threaded into the step).
- HubSpot schema moved out of the 'use server' module; client embed reads
  NEXT_PUBLIC_* directly so Next inlines it in the browser bundle.

Cleanup:
- Removed dead Zustand WebGL-tunnel store.
- Deduped Shopify install via installBundle and fluid/flowmap pointer input
  via a shared usePointerInput hook.
- Split ast-transforms.ts into helpers/remove-ops/add-ops/index (barrel);
  setup-project.ts under 1k lines.
- Honest bundle-key typing (BundleId + getBundle); dropped dishonestly
  optional types (CartLineItem.id, CartMeta, discriminated CanvasContextValue).
- Server-side integration env reads routed through the typed @/lib/env.
