+++ lib/webgl/store.ts
@@ -1,31 +1,20 @@
 import tunnel from 'tunnel-rat'
-import { create } from 'zustand'
 
 type WebGLTunnelInstance = ReturnType<typeof tunnel>
 
 let webGLTunnelSingleton: WebGLTunnelInstance | null = null
 let domTunnelSingleton: WebGLTunnelInstance | null = null
 
-function getWebGLTunnel(): WebGLTunnelInstance {
+export function getWebGLTunnel(): WebGLTunnelInstance {
   if (!webGLTunnelSingleton) {
     webGLTunnelSingleton = tunnel()
   }
   return webGLTunnelSingleton
 }
 
-function getDOMTunnel(): WebGLTunnelInstance {
+export function getDOMTunnel(): WebGLTunnelInstance {
   if (!domTunnelSingleton) {
     domTunnelSingleton = tunnel()
   }
   return domTunnelSingleton
 }
-
-type WebGLStore = {
-  getWebGLTunnel: () => WebGLTunnelInstance
-  getDOMTunnel: () => WebGLTunnelInstance
-}
-
-export const useWebGLStore = create<WebGLStore>(() => ({
-  getWebGLTunnel,
-  getDOMTunnel,
-}))
